stable release v1.0.1
